# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateProjectRequest:
    properties:
      description:
        example: My first food delivery app
        type: string
      name:
        example: Foodie
        type: string
    required:
    - name
    type: object
...
  handlers.UpdateProjectStatusRequest:
    properties:
      isActive:
        type: boolean
    type: object
  models.Collection:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        description: e.g., "users", "products"
        type: string
      projectId:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Project:
    properties:
      apiKey:
        description: Permanent key for client-side frontend apps
        type: string
      collections:
        items:
          $ref: '#/definitions/models.Collection'
        type: array
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isActive:
        description: Can be disabled instead of deleted
        type: boolean
      name:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A multi-tenant JSONB document store for students.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VelocityDBGo API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      description: Retrieve all projects belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a logical container for an application
      parameters:
      - description: Project parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
...
  /api/projects/{projectId}/status:
    patch:
      consumes:
      - application/json
      description: Toggle whether a project is active or disabled (soft delete). Disabled
        projects cannot accept Data API queries.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Status parameter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProjectStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enable/Disable a project
      tags:
      - Projects
securityDefinitions:
  ApiKeyAuth:
    description: Public API Key for a specific Project.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"